Feat: buddy matching api #2

+++ backend/moa/src/main/java/com/ssafy/moa/api/service/BuddyService.java
@@ -5,5 +5,5 @@ import com.ssafy.moa.api.dto.BuddyDto.*;
 public interface BuddyService {
     Long saveKoreanBuddyInfo(KoreanBuddyPostRequest koreanBuddyPostRequest);
     Long saveForeignerBuddyInfo(ForeignerBuddyPostRequest foreignerBuddyPostRequest);
-    Integer findMatchingBuddy(BuddyMatchingRequest buddyMatchingRequest);
+    Long findMatchingBuddy(BuddyMatchingRequest buddyMatchingRequest);
 }
